Use one file per command
